mw/com: Use explicit static_cast to avoid implicit conversions
issue: SWP-218390
GIT_ORIGIN_SPP_REV_ID: f5d0acfe23534b59d57c75c3745cfa24abb6a176

+++ score/mw/com/impl/bindings/lola/application_id_pid_mapping_entry.cpp
@@ -43,7 +43,7 @@ ApplicationIdPidMappingEntry::key_type ApplicationIdPidMappingEntry::CreateKey(M
 {
     ApplicationIdPidMappingEntry::key_type result = static_cast<std::uint16_t>(status);
     result = result << 32U;
-    result |= application_id;
+    result |= static_cast<std::uint64_t>(application_id);
     return result;
 }
